feat: #991 move tooltip add for monitoring liq auction
