--- apps/trading/components/trading-mode-tooltip/index.tsx
@@ -1,1 +1,0 @@
-export * from './trading-mode-tooltip';
