+++ apps/trading/pages/markets/trade-grid.tsx
@@ -1,4 +1,8 @@
-import { DealTicketContainer } from '@vegaprotocol/deal-ticket';
+import {
+  compileGridData,
+  DealTicketContainer,
+  TradingModeTooltip,
+} from '@vegaprotocol/deal-ticket';
 import { MarketInfoContainer } from '@vegaprotocol/market-info';
 import { OrderbookContainer } from '@vegaprotocol/market-depth';
 import { ColumnKind, SelectMarketPopover } from '@vegaprotocol/market-list';
@@ -39,7 +43,6 @@ import {
   MarketTradingMode,
   MarketTradingModeMapping,
 } from '@vegaprotocol/types';
-import { TradingModeTooltip } from '../../components/trading-mode-tooltip';
 import { Header, HeaderStat } from '../../components/header';
 
 const TradingViews = {
@@ -173,9 +176,7 @@ export const TradeMarketHeader = ({
         description={
           <TradingModeTooltip
             market={market}
-            onSelect={(marketId: string) => {
-              onSelect(marketId);
-            }}
+            compiledGrid={compileGridData(market, onSelect)}
           />
         }
       >
