+++ libs/deal-ticket/src/components/deal-ticket/deal-ticket.spec.tsx
@@ -27,6 +27,7 @@ const market: DealTicketMarketFragment = {
           id: 'asset-id',
           name: 'asset-name',
           symbol: 'asset-symbol',
+          decimals: 2,
         },
       },
     },
