+++ libs/deal-ticket/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './deal-ticket';
 export * from './deal-ticket-validation';
+export * from './trading-mode-tooltip';
